add implementation for font8

+++ liboled/c/include/sdcc/oled_font8.h
@@ -38,6 +38,7 @@ struct oled_font8_context
     void *font;
 };
 
+
 /**
  * Initialise the font8 context structure
  *
@@ -46,18 +47,32 @@ struct oled_font8_context
  * @param font     pointer to font. font data is expected indexed from 0
  * @param font_width  width of font
  */
-extern void oled_font8_init(struct oled_font8_context *context, uint8_t *buffer, uint8_t *font, uint8_t font_width) __z88dk_callee;
+extern void oled_font8_init(struct oled_font8_context *context, uint8_t *buffer, uint8_t *font, uint16_t font_width) __z88dk_callee;
+
 
 /**
- * Set a different font
+ * Set a font to print with.
+ *
+ * Fonts are expected to be 8bits per column
  *
  * @param context  valid font8 context
  * @param font     pointer to font. font data is expected indexed from 0
  * @param font_width  width of font
  */
-extern void oled_font8_set_font(struct oled_font8_context *context, uint8_t *font, uint8_t font_width) __z88dk_callee;
+extern void oled_font8_set_font(struct oled_font8_context *context, uint8_t *font, uint16_t font_width) __z88dk_callee;
 
+
+/**
+ * Set row and column position for printing
+ *
+ * @param context  valid font8 context
+ * @param row      row number (0-3 - where 0 is the top)
+ * @param column   column (0-127)
+ */
 extern void oled_font8_set_rc(struct oled_font8_context *context, uint8_t row, uint8_t column) __z88dk_callee;
+
+
+
 extern void oled_font8_set_xy(struct oled_font8_context *context, uint8_t x, uint8_t y) __z88dk_callee;
 
 extern void oled_font8_putc(struct oled_font8_context *context, uint8_t ch)  __z88dk_callee;
